+++ partials/footers.php
@@ -11,5 +11,5 @@
 
 <footer>
 	<p>Dulcet Games Inc. © 2016 - 2018. All rights reserved.</p><br>
-	<p class="footer_link"><a href="index.php">HOME</a> | <a href="terms.php">TERMS</a> | <a href="#">SHOP</a> | <a href="about.php">ABOUT</a> | <a href="contact.php">CONTACT</a> | <a href="parents.php">FOR PARENTS</a></p>
+	<p class="footer_link"><a href="index.php">HOME</a> | <a href="terms.php">TERMS</a> | <a href="shop.php">SHOP</a> | <a href="about.php">ABOUT</a> | <a href="contact.php">CONTACT</a> | <a href="parents.php">FOR PARENTS</a></p>
 </footer>